+++ prow/jobs/control-plane/control-plane-reconciler-integration.yaml
@@ -37,7 +37,7 @@ presubmits: # runs on PRs
           base_ref: main
       spec:
         containers:
-          - image: "eu.gcr.io/kyma-project/test-infra/kyma-integration:v20211022-7a608efc"
+          - image: "eu.gcr.io/kyma-project/test-infra/kyma-integration:v20211020-638b6f95"
             securityContext:
               privileged: true
             command:
